backend/requirements.txt
fastapi
uvicorn
pandas
numpy
duckdb
transformers
torch
sentence-transformers
faiss-cpu
spacy
